refactor(board-webmvc): use base-package as `nettee.board.driving.web`
