+++ services/board/driving/web-mvc/src/main/java/nettee/board/driving/web/dto/BoardCommandDto.java
@@ -1,4 +1,4 @@
-package nettee.board.web.dto;
+package nettee.board.driving.web.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
