+++ services/board/driving/web-mvc/src/main/java/nettee/board/driving/web/mapper/BoardDtoMapper.java
@@ -1,8 +1,8 @@
-package nettee.board.web.mapper;
+package nettee.board.driving.web.mapper;
 
 import nettee.board.domain.Board;
 import nettee.board.domain.type.BoardStatus;
-import nettee.board.web.dto.BoardCommandDto.BoardCreateCommand;
+import nettee.board.driving.web.dto.BoardCommandDto.BoardCreateCommand;
 import org.mapstruct.Mapper;
 
 @Mapper(componentModel = "spring")
